web_app/requirements.txt
Django==2.2.9
django-templated-mail==1.1.1
django-vote==2.1.7
djangorestframework==3.11.0
djoser==2.0.3
pytz==2019.3
sqlparse==0.3.0
gunicorn